app/requirements.txt
astroid==1.6.2
cachetools==2.0.1
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.6
django-rest-swagger==2.1.2
django-sslserver==0.20
djangorestframework==3.7.1
google-api-python-client==1.6.5
google-auth==1.3.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
httplib2==0.10.3
idna==2.6
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
oauth2client==4.1.2
oauthlib==2.0.6
openapi-codec==1.3.1
psycopg2==2.7.3.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pylint==1.8.3
python-magic==0.4.15
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
simplejson==3.13.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
wrapt==1.10.11
